Refactor Discord service imports and add Channel and Embed types
