+++ src/api/services/matches-service.ts
@@ -9,20 +9,22 @@ import {
   markMatchComplete,
   updateMatchProcessed,
 } from "../../db/commands";
-import {
-  createLiveScoreCard,
-  createMatchAnalysisEmbed,
-  deleteMatchCards,
-  runEloUpdate,
-  sendMatchFinishNotification,
-  updateVoiceChannelStatus,
-} from "./discord-service";
 import { FaceitService } from "./faceit-service";
 import {
   aggregateEnemyFactionData,
   formatMapData,
   getScoreStatusText,
 } from "../../utils/faceitHelper";
+import { updateChannelStatus } from "./discord/channel-service";
+import {
+  createLiveScoreCard,
+  deleteMatchCards,
+} from "./discord/live-game-service";
+import {
+  sendMatchEndSummary,
+  sendMatchroomAnalysisEmbed,
+} from "./discord/embed-service";
+import { runEloUpdate } from "./discord/member-service";
 
 let workers: Record<string, Worker> = {};
 
@@ -46,7 +48,10 @@ export const startMatch = async (matchId: string) => {
   // If the players are in a voice channel. Create a JS Worker to update the live score in the status of the channel.
   if (match?.voiceChannelId) {
     const scoreStatus = await getScoreStatusText(match.mapName);
-    await updateVoiceChannelStatus(match.voiceChannelId, scoreStatus);
+    await updateChannelStatus({
+      id: match.voiceChannelId,
+      status: scoreStatus,
+    });
   }
 
   await createLiveScoreCard(match);
@@ -87,11 +92,11 @@ export const endMatch = async (matchId: string) => {
 
   // deletes record from DB.
   await markMatchComplete(matchId);
-  await sendMatchFinishNotification(match);
+  await sendMatchEndSummary(match);
   await runEloUpdate(match.trackedTeam.trackedPlayers);
 
   if (match?.voiceChannelId) {
-    await updateVoiceChannelStatus(match.voiceChannelId, "");
+    await updateChannelStatus({ id: match.voiceChannelId, status: "" });
   }
 
   await deleteMatchCards(matchId);
@@ -118,7 +123,7 @@ export const cancelMatch = async (matchId: string) => {
   }
 
   if (match?.voiceChannelId) {
-    await updateVoiceChannelStatus(match.voiceChannelId, "");
+    await updateChannelStatus({ id: match.voiceChannelId, status: "" });
   }
 
   await deleteMatchCards(matchId);
@@ -166,5 +171,5 @@ export const getMatchAnalysis = async (matchId: string): Promise<any> => {
     matchroomPlayers.enemyFaction.length
   );
 
-  createMatchAnalysisEmbed(matchId, matchroomPlayers, formattedMapData);
+  sendMatchroomAnalysisEmbed(matchId, matchroomPlayers, formattedMapData);
 };
