+++ src/bot/main.ts
@@ -4,7 +4,6 @@ import { config } from "../config";
 import { commands, commandsMap } from "./commands";
 import "./events/welcome";
 import "./events/manage-voice-states";
-import "./events/ai";
 import "./events/leaver";
 import "./events/register";
