+++ src/types/Discord/Channel.ts
@@ -1,0 +1,16 @@
+export type Channel = {
+  id: string;
+  name: string;
+  status?: string;
+};
+
+export type CreateChannel = {
+  channelName: string;
+  categoryId: string;
+  locked?: boolean;
+};
+
+export type UpdateChannelStatus = {
+  id: string;
+  status: string;
+};
